+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 14.1.0 Fixes
 
 - `Button` Rearrange button layout to fix notificationBadge alignment. ([#1319](https://github.com/infor-design/enterprise-ng/issues/1319))
+- `Datepicker` Fix a bug where datepicker is displaying NaN in french format. ([#1273](https://github.com/infor-design/enterprise-ng/issues/1273))
 
 ## 14.0.0